archive/v2_backup/mlops_requirements.txt
# MLOps Enhanced Requirements for Magnus Carlsen Training

# Core ML libraries (already installed)
torch>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.62.0

# Chess libraries (already installed)
python-chess>=1.999
stockfish>=3.28.0

# MLOps and Experiment Tracking
mlflow>=2.8.0
evidently>=0.4.0

# Optional: Enhanced monitoring and visualization
optuna>=3.0.0          # Hyperparameter optimization
wandb>=0.15.0          # Alternative to MLflow
tensorboard>=2.13.0    # TensorBoard logging
prometheus-client>=0.17.0  # Metrics collection
plotly>=5.15.0         # Interactive plots

# Data versioning and pipeline
dvc>=3.0.0             # Data version control
great-expectations>=0.17.0  # Data validation

# Model serving (for deployment)
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
